+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 -   dropped internal mbtiles package in favor of github.com/brendan-ward/mbtiles-go,
     which wraps the SQlite-specific go package `crawshaw.io/sqlite`
 
+### Command-line interface
+
+-   added support for watching filesystem for changes to tilesets using `--enable-fs-watch` option
+
 ### Breaking changes
 
 -   removes ArcGIS API layer info at the service root and layers endpoint (#116);